refactored contacts
